changed input tensor to float32

+++ classifier/classifier.py
@@ -91,7 +91,8 @@ while True:
         if path.exists(WAV_PATH + row[1]):
             predict_x = extract_features_only(WAV_PATH + row[1])
             predict_x = np.expand_dims(predict_x, axis=0) # expand from 3D to 4D
-            interpreter.set_tensor(input_details[0]['index'], predict_x.astype(np.uint8))
+            #interpreter.set_tensor(input_details[0]['index'], predict_x.astype(np.uint8))
+            interpreter.set_tensor(input_details[0]['index'], predict_x.astype(np.float32))
             #input_details = interpreter.get_input_details()[0] # for one input data
             #tensor_index = input_details['index'] # tensor index in the interpreter
             #input_tensor = interpreter.tensor(tensor_index)()[0]
